+++ www/pt-wc-q/js/extended/events.js
@@ -1,5 +1,5 @@
 import {eventsPageState} from '../events_page_state.js'
-import {crud_assembly} from '../../../components/adminui-custom/components/adminui-crud.js';
+import {crud_assembly} from '../../../components/adminui-custom/components/adminui-crud-custom.js';
 
 let events_exteneded = {
     ...eventsPageState,
